Moved files around (see comment)
- Created a LargeButton Component instead of creating individual buttons
- Renamed animations to assets
- Fixed a dependency error with react-router-dom
